Revert "Move string makers"
This reverts commit 406dc07c413ae27b9a0e92c394f7be496820212b.
